requirements.txt
django
djangorestframework
markdown
django-filter
django-bootstrap3
uuid
celery
flower

